adding test setup for API
