+++ .travis.yml
@@ -1,12 +1,15 @@
 language: node_js
 node_js:
   - "10"
+services:
+  - mongodb
 cache:
   yarn: true
 install:
   - yarn
 script:
   - yarn workspace api build
+  - yarn workspace api test
   - yarn workspace client typecheck
   - yarn workspace client build
   - yarn workspace client test
