+++ api/package.json
@@ -45,17 +45,18 @@
     "vision": "^5.4.2"
   },
   "devDependencies": {
-    "jest": "^22.4.0",
+    "@types/jest": "^23.3.11",
+    "jest": "^23.6.0",
     "micro-dev": "^3.0.0",
-    "nodemon": "^1.18.7",
+    "nodemon": "^1.18.9",
+    "ts-jest": "^23.10.5",
     "ts-node": "^7.0.1",
     "typedoc": "^0.12.0",
     "yesno-http": "^0.0.5"
   },
   "jest": {
-    "transform": {
-      "^.+\\.jsx?$": "babel-jest"
-    }
+    "preset": "ts-jest",
+    "testEnvironment": "node"
   },
   "nodemonConfig": {
     "watch": [
