+++ api/tsconfig.json
@@ -1,7 +1,6 @@
 {
   "compilerOptions": {
-    "typeRoots": ["types", "node_modules/@types"],
-    "lib": ["es2016"],
+    "lib": ["es2016", "dom"],
     "module": "commonjs",
     "target": "es6",
     "moduleResolution": "node",
@@ -13,5 +12,5 @@
     }
   },
   "include": ["types/*", "src/**/*"],
-  "exclude": ["dist", "docs", "test"]
+  "exclude": ["dist", "docs"]
 }
